lnTerm separated
